decouple b plus tree debugging tool
